+++ app/src/main/AndroidManifest.xml
@@ -104,7 +104,7 @@
             android:name=".ui.JoinActivity"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".ui.ObservationActivity"
+            android:name=".ui.observations.ObservationActivity"
             android:configChanges="orientation|keyboardHidden|screenSize"
             android:parentActivityName=".ui.MainActivity"
             android:screenOrientation="portrait">
@@ -113,7 +113,7 @@
                 android:value=".ui.MainActivity" />
         </activity>
         <activity
-            android:name=".ui.AddObservationActivity"
+            android:name=".ui.observations.AddObservationActivity"
             android:parentActivityName=".ui.MainActivity"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize">
@@ -122,7 +122,7 @@
                 android:value=".ui.MainActivity" />
         </activity>
         <activity
-            android:name=".ui.ProjectActivity"
+            android:name=".ui.projects.ProjectActivity"
             android:parentActivityName=".ui.MainActivity"
             android:screenOrientation="portrait">
             <meta-data
@@ -130,7 +130,7 @@
                 android:value=".ui.MainActivity" />
         </activity>
         <activity
-            android:name=".ui.UsersDetailActivity"
+            android:name=".ui.communities.UsersDetailActivity"
             android:parentActivityName=".ui.MainActivity"
             android:screenOrientation="portrait">
             <meta-data
@@ -168,9 +168,9 @@
             android:name="com.google.firebase.messaging.default_notification_color"
             android:resource="@color/colorAccent" />
 
-        <activity android:name=".ui.IdeaDetailsActivity" />
-        <activity android:name=".ui.AddDesignIdeaActivity" />
-        <activity android:name=".ui.AddProjectActivity" />
+        <activity android:name=".ui.ideas.IdeaDetailsActivity" />
+        <activity android:name=".ui.ideas.AddDesignIdeaActivity" />
+        <activity android:name=".ui.projects.AddProjectActivity" />
         <activity android:name=".ui.ImagePicker" />
         <activity android:name=".ui.UserProfileSettings" />
         <activity android:name=".ui.SettingsActivity"></activity>
