+++ app/src/main/java/org/naturenet/ui/observations/AddObservationImageAdapter.java
@@ -1,4 +1,4 @@
-package org.naturenet.ui;
+package org.naturenet.ui.observations;
 
 import android.content.Context;
 import android.net.Uri;
