+++ app/src/main/res/layout/activity_add_observation.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:background="@color/colorTheme"
-    tools:context="org.naturenet.ui.AddObservationActivity">
+    tools:context="org.naturenet.ui.observations.AddObservationActivity">
 
     <include
         layout="@layout/app_bar_add_observation"
