+++ app/src/main/res/layout/activity_add_project.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.naturenet.ui.AddProjectActivity">
+    tools:context="org.naturenet.ui.projects.AddProjectActivity">
 
 
     <include
