+++ app/src/main/res/layout/activity_idea_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.naturenet.ui.IdeaDetailsActivity">
+    tools:context="org.naturenet.ui.ideas.IdeaDetailsActivity">
 
 
     <include
