+++ app/src/main/res/layout/activity_observation.xml
@@ -5,7 +5,7 @@
     android:layout_height="fill_parent"
     android:fitsSystemWindows="true"
     android:background="@color/colorTheme"
-    tools:context="org.naturenet.ui.ObservationActivity">
+    tools:context="org.naturenet.ui.observations.ObservationActivity">
 
     <include
         layout="@layout/app_bar_observation"
