+++ app/src/main/res/layout/app_bar_add_design_idea.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="org.naturenet.ui.AddDesignIdeaActivity">
+    tools:context="org.naturenet.ui.ideas.AddDesignIdeaActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
