+++ app/src/main/res/layout/app_bar_add_observation.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="org.naturenet.ui.AddObservationActivity">
+    tools:context="org.naturenet.ui.observations.AddObservationActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
