+++ app/src/main/res/layout/app_bar_project.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="org.naturenet.ui.ProjectActivity">
+    tools:context="org.naturenet.ui.projects.ProjectActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
