+++ app/src/main/res/layout/app_bar_user.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="org.naturenet.ui.UsersDetailActivity">
+    tools:context="org.naturenet.ui.communities.UsersDetailActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
