+++ app/src/main/res/layout/content_observation.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@android:color/white"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="org.naturenet.ui.ObservationActivity"
+    tools:context="org.naturenet.ui.observations.ObservationActivity"
     tools:showIn="@layout/app_bar_observation">
 
     <FrameLayout
