+++ app/src/main/res/layout/content_project.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@android:color/white"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="org.naturenet.ui.ProjectActivity"
+    tools:context="org.naturenet.ui.projects.ProjectActivity"
     tools:showIn="@layout/app_bar_project">
 
     <FrameLayout
