+++ app/src/main/res/layout/fragment_add_design_idea.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.naturenet.ui.AddDesignIdeaFragment">
+    tools:context="org.naturenet.ui.ideas.AddDesignIdeaFragment">
 
     <TextView
         android:id="@+id/designIdeaDescription"
