+++ app/src/main/res/layout/fragment_add_observation.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.naturenet.ui.AddObservationFragment">
+    tools:context="org.naturenet.ui.observations.AddObservationFragment">
 
     <RelativeLayout
         android:id="@+id/projects_layout"
