+++ app/src/main/res/layout/fragment_communities.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="org.naturenet.ui.CommunitiesFragment">
+             tools:context="org.naturenet.ui.communities.CommunitiesFragment">
 
     <EditText
         android:id="@+id/searchText"
