+++ app/src/main/res/layout/fragment_idea_details.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.naturenet.ui.IdeaDetailsFragment">
+    tools:context="org.naturenet.ui.ideas.IdeaDetailsFragment">
 
 
     <LinearLayout
