+++ app/src/main/res/layout/fragment_ideas.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="org.naturenet.ui.IdeasFragment">
+             tools:context="org.naturenet.ui.ideas.IdeasFragment">
 
     <!--<include
         layout="@layout/floating_buttons"
