+++ app/src/main/res/layout/fragment_observation.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:layout_margin="10dp"
     android:id="@+id/mainLayout"
-    tools:context="org.naturenet.ui.ObservationFragment">
+    tools:context="org.naturenet.ui.observations.ObservationFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
