+++ app/src/main/res/layout/fragment_project_detail.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.naturenet.ui.ProjectDetailFragment">
+    tools:context="org.naturenet.ui.projects.ProjectDetailFragment">
 
     <RelativeLayout
         android:id="@+id/project_rl"
